Decompose recipe display components/pages into smaller modules. Improved code consstency with inline className assignments
