+++ components/Searchbar.jsx
@@ -81,8 +81,9 @@ export const Searchbar = () => {
               transition={{ duration: 0.7 }}
             >
               <input
-                className={`form-control relative flex-auto min-w-0 w-full px-4 py-2 text-base font-normal text-gray-700 bg-white bg-clip-padding 
-                  border border-gray-300 border-solid rounded-lg transition ease-in-out m-0 focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none
+                className={`form-control relative flex-auto min-w-0 w-full text-base font-normal text-gray-700 bg-white bg-clip-padding 
+                  focus:text-gray-700 focus:bg-white focus:border-blue-600 focus:outline-none
+                  border border-gray-300 border-solid rounded-lg transition ease-in-out m-0 px-4 py-2
                   ${error ? 'error-border' : ''}`} 
                 type="text"
                 aria-label="Search" 
@@ -94,15 +95,17 @@ export const Searchbar = () => {
               />
               <button 
                 type="submit" 
-                className="search-main relative px-4 py-1 ml-4 border-2 border-blue-400 text-blue-400 md:font-medium text-xs leading-tight uppercase 
-                  rounded-full hover:bg-black hover:bg-opacity-5 focus:outline-none focus:ring-0 transition duration-150 ease-in-out"
+                className="search-main relative border-2 border-blue-400 text-blue-400 md:font-medium text-xs leading-tight uppercase 
+                  rounded-full hover:bg-black hover:bg-opacity-5 focus:outline-none focus:ring-0 transition duration-150 ease-in-out px-4 py-1 ml-4"
                 aria-label="Submit search"
                 aria-controls="submit-search"
                 aria-describedby="main-search-submit"
                 role="button"
                 tabIndex="0"
               >
-                <span id="search-icon" className="input-group-text flex flex items-center md:px-3.5 py-1 cursor-pointer text-base font-normal text-gray-600 dark:text-gray-200 text-center whitespace-nowrap rounded">
+                <span id="search-icon" className="input-group-text flex flex items-center rounded cursor-pointer 
+                  text-base font-normal text-gray-600 dark:text-gray-200 text-center whitespace-nowrap md:px-3.5 py-1"
+                >
                   <svg className="w-4" aria-hidden="true" focusable="false" data-prefix="fas" data-icon="search" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
                     <path fill="currentColor" d="M505 442.7L405.3 343c-4.5-4.5-10.6-7-17-7H372c27.6-35.3 44-79.7 44-128C416 93.1 322.9 0 208 0S0 93.1 0 208s93.1 208 208 208c48.3 0 92.7-16.4 128-44v16.3c0 6.4 2.5 12.5 7 17l99.7 99.7c9.4 9.4 24.6 9.4 33.9 0l28.3-28.3c9.4-9.4 9.4-24.6.1-34zM208 336c-70.7 0-128-57.2-128-128 0-70.7 57.2-128 128-128 70.7 0 128 57.2 128 128 0 70.7-57.2 128-128 128z"></path>
                   </svg>
