+++ data/recipeDb.js
@@ -6,9 +6,9 @@ export const recipes = [
     "name": "One Pot Creamy Pasta",
     "images": ["/pasta.png"],
     "types": ["pasta"],
+    "time": "25 minutes",
     "vegetarian": true,
     "vegan": false,
-    "time": "25 minutes",
     "ingredients": [
       "4 garlic cloves, peeled & sliced",
       "7 oz (1 jar) sun dried tomatoes, drained",
@@ -47,9 +47,9 @@ export const recipes = [
     "name": "Buffalo Chicken Enchiladas",
     "images": ["/chicken.png", "/texmex.png"],
     "types": ["chicken", "texMex"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "30 minutes",
     "ingredients": [
       "3 tbsp butter, melted",
       "4 cups shredded rotisserie chicken",
@@ -86,9 +86,9 @@ export const recipes = [
     "name": "Chili Mac and Cheese",
     "images": ["/pasta.png", "/beef.png"],
     "types": ["pasta", "beef"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "",
     "ingredients": [
       "1 tbsp olive oil",
       "2 cloves garlic, minced",
@@ -127,10 +127,10 @@ export const recipes = [
     "id": 3,
     "name": "Bean and Cheese Burritos",
     "images": ["/vegetarian.png", "/texmex.png"],
-    "types": ["vegetarian, texMex"],
+    "types": ["vegetarian", "texMex"],
+    "time": "25 minutes",
     "vegetarian": true,
     "vegan": false,
-    "time": "25 minutes",
     "ingredients": [
       "1 can (16 oz) vegetarian refried beans",
       "1 cup salsa",
@@ -160,9 +160,9 @@ export const recipes = [
     "name": "Sunset Thai Pasta",
     "images": ["/pasta.png"],
     "types": ["pasta"],
+    "time": "30 minutes",
     "vegetarian": true,
     "vegan": false,
-    "time": "30 minutes",
     "ingredients": [
       "12 oz dry fettuccine",
       "4 1/2 cups of vegetable broth",
@@ -204,9 +204,9 @@ export const recipes = [
     "name": "Spicy Shepherd's Pie",
     "images": ["/beef.png"],
     "types": ["beef"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "30 minutes",
     "ingredients": [
       "3 cups mashed potato flakes",
       "1 lb ground beef",
@@ -241,9 +241,9 @@ export const recipes = [
     "name": "Pork Fried Rice", 
     "images": ["/pork.png"],
     "types": ["pork"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "30 minutes",
     "ingredients": [
       "4 cups basmati rice",
       "8 cups water",
@@ -281,9 +281,9 @@ export const recipes = [
     "name": "Broccoli Cheddar Soup", 
     "images": ["/soup.png"],
     "types": ["soup"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "30 minutes",
     "ingredients": [
       "1/4 cup chopped onion",
       "1/4 cup butter, cubed",
@@ -316,9 +316,9 @@ export const recipes = [
     "name": "Pasta Fagioli Soup", 
     "images": ["/soup.png", "/pasta.png"],
     "types": ["soup", "pasta"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "30 minutes",
     "ingredients": [
       "1 lb ground Italian sausage",
       "1 small onion, chopped",
@@ -357,9 +357,9 @@ export const recipes = [
     "name": "Southwestern Chicken Salad", 
     "images": ["/salad.png", "/chicken.png"],
     "types": ["salad", "chicken"],
+    "time": "30 minutes",
     "vegetarian": false,
     "vegan": false,
-    "time": "",
     "ingredients": [
       "1 lb chicken breast",
       "1 tbsp",
@@ -398,10 +398,10 @@ export const recipes = [
     "id": 10,
     "name": "Creamy Chicken Skillet", 
     "images": ["/chicken.png"],
+    "time": "20 minutes",
     "types": ["chicken"],
     "vegetarian": false,
     "vegan": false,
-    "time": "20 minutes",
     "ingredients": [
       "1 lb chicken cutlets",
       "¼ tsp salt, divided",
@@ -437,10 +437,10 @@ export const recipes = [
     "id": 11,
     "name": "Coconut Chickpea Curry", 
     "images": ["/vegetarian.png", "/vegan.png"],
+    "time": "20 minutes",
     "types": ["vegetarian"],
     "vegetarian": true,
     "vegan": true,
-    "time": "20 minutes",
     "ingredients": [
       "2 tsp avocado oil or canola oil",
       "1 cup chopped onion",
@@ -472,10 +472,10 @@ export const recipes = [
     "id": 12,
     "name": "Tomato Salmon Skillet", 
     "images": ["/seafood.png"],
+    "time": "20 minutes",
     "types": ["seafood"],
     "vegetarian": false,
     "vegan": false,
-    "time": "20 minutes",
     "ingredients": [
       "1¼ lb salmon, skinned and cut into 4 portions",
       "¼ tsp salt",
@@ -513,10 +513,10 @@ export const recipes = [
     "id": 13,
     "name": "Super Shrimp Scampi", 
     "images": ["/seafood.png"],
+    "time": "20 minutes",
     "types": ["seafood"],
     "vegetarian": false,
     "vegan": false,
-    "time": "20 minutes",
     "ingredients": [
       "1½ lb fresh or frozen shrimp in shells",
       "6 oz whole-wheat or plain linguine",
@@ -548,10 +548,10 @@ export const recipes = [
     "id": 14,
     "name": "Skillet Ravioli Lasagna", 
     "images": ["/pasta.png", "/beef.png"],
+    "time": "20 minutes",
     "types": ["pasta"],
     "vegetarian": false,
     "vegan": false,
-    "time": "20 minutes",
     "ingredients": [
       "1 (24 oz) package frozen ravioli",
       "1 lb lean ground beef",
@@ -584,10 +584,10 @@ export const recipes = [
     "id": 15,
     "name": "Korean Beef and Rice", 
     "images": ["/beef.png"],
+    "time": "25 minutes",
     "types": ["beef"],
     "vegetarian": false,
     "vegan": false,
-    "time": "20 minutes",
     "ingredients": [
       "1 lb lean ground beef",
       "3 garlic cloves, minced",
