+++ styles/globals.css
@@ -87,6 +87,21 @@ button:focus {
   }
 }
 
+.fade-in {
+  animation-name: fade-in;
+  animation-duration: 1.5s;
+  animation-fill-mode: both;
+}
+
+@keyframes fade-in {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
 .jello {
   animation: jello-jiggle 2.2s ease-in;
 }
